background-job-processing-system/requirements.txt
Django>=5.2.3
celery>=5.3.6
django-celery-results>=2.5.1
django-celery-beat>=2.6.0
djangorestframework>=3.15.1
redis>=5.0.4
